# Optimise TextEdit base operations