--- a/scene/gui/control.cpp
+++ b/scene/gui/control.cpp
@@ -2597,7 +2597,7 @@ bool Control::is_text_field() const {
 	return false;
 }
 
-Array Control::structured_text_parser(StructuredTextParser p_theme_type, const Array &p_args, const String p_text) const {
+Array Control::structured_text_parser(StructuredTextParser p_theme_type, const Array &p_args, const String &p_text) const {
 	Array ret;
 	switch (p_theme_type) {
 		case STRUCTURED_TEXT_URI: {