Significantly increased the performance of base `TextEdit` operations, not sure we can get much more out of these now, unless we move away from using `vector`. The remaining costly operations seem to now be tied up in `TextServer`.

Some benchmarks during my testing:

- Load time for a `21k` script as been reduced from `10.5s` to `1.3s`. 
- Deleting `21k` lines has been decreased from `95s` to `90ms`.
- Inserting `21k` lines has been reduced by around `300ms`

As for the changes, previously for both insert and remove we were going line by line causing a `vector` resize each time, now it operates on a range so we only need to resize and shuffle the `vector` once. 

During load we called `invalidate_all` which updated line by line, however with the cached values not matching the line, we would excessively call  `_calculate_line_height` and `_calculate_max_line_width` to check if they are valid, which would be the case until the last line had been updated. This has been updated to clear both values where we can.

Made some smaller adjustments to reduce string operations, such as `_base_get_text` now uses the string builder.

closes #42496