@Zireael07 2021-12-08
> Deleting 21k lines has been decreased from 95s to 90ms.

Typo? Or indeed a 1000x speedup?!