@Thed13 2021-12-09
> > Deleting 21k lines has been decreased from 95s to 90ms.
> 
> Typo? Or indeed a 1000x speedup?!

Yea, how did he do that ?!, without any compromises ?