@Paulb23 2021-12-09
> Typo? Or indeed a 1000x speedup?!
 
Not a typo, in `ms` it was `95662ms`.

The old code removed from the front of the `vector` for the amount of times we needed, removing from the font means we have to push all the other elements back by `1`. So for `20k` lines it would move `19,999` lines back `1` place, resize to remove a line, invalidate cache, move `19,998` back `1` place, resize to remove a line, invalidate cache and so on. Something around `O(n^2)`.

New code moves only the lines we need in one go, resizes and cache update. On a delete all, this is pretty close to `O(1)` as we only care about the first line (Technically `O(n)`).